requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arabic-reshaper==2.1.3
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
budou==0.9.8
certifi==2026.2.25
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
datasets==4.6.1
diffimg==0.2.3
dill==0.4.0
docopt==0.6.2
filelock==3.24.3
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
fugashi==1.5.2
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
h11==0.16.0
hf-xet==1.3.2
html2image==2.0.7
html5lib==1.0.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
ipadic==1.0.0
itsdangerous==2.2.0
jaconv==0.5.0
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
loguru==0.7.3
-e git+https://github.com/kha-white/manga-ocr.git@847e0939fcb391ad63584dcf50ba6c6533ec9ee8#egg=manga_ocr
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mecab-python3==1.0.12
ml_dtypes==0.5.4
mokuro==0.2.4
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
natsort==8.4.0
networkx==3.6.1
numpy==2.4.2
onnx==1.20.1
onnx-simplifier==0.5.0
onnxsim==0.4.36
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.1
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pyperclip==1.11.0
pytest==9.0.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rsa==4.9.1
ruff==0.15.4
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.53.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.13
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
stringzilla==4.6.0
sympy==1.14.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2026.2.24
tinysegmenter3==0.1.0
tokenizers==0.22.2
torch==2.10.0
torchinfo==1.8.0
torchsummary==1.5.1
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
trdg==1.8.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unidic-lite==1.0.8
uritemplate==4.2.0
urllib3==2.6.3
wandb==0.25.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.6
wikipedia==1.4.0
xxhash==3.6.0
yarl==1.22.0
yattag==1.16.1
